requirements.txt
# Google Cloud SDK dependencies for Video Reasoning Prototype
google-cloud-aiplatform>=1.38.0
google-cloud-videointelligence>=2.11.0
google-cloud-storage>=2.10.0
vertexai>=1.0.0
google-generativeai>=0.8.0

# Additional utilities
python-dotenv>=1.0.0
